requirements.txt
flask
faiss-cpu
sentence-transformers
tiktoken
requests